requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.6
django-jazzmin==3.0.1
django-widget-tweaks==1.5.0
pillow==11.3.0
PyMySQL==1.1.2
python-decouple==3.8
setuptools==80.9.0
sqlparse==0.5.3
whitenoise==6.11.0
